# public/openapi.yaml
openapi: 3.0.0
info:
  title: Hermes Agent Manager API
  version: 1.0.0
  description: API for managing remote Docker agents, servers, and configurations.

servers:
  - url: http://localhost:3000
    description: Local development server

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY

security:
  - ApiKeyAuth: []

paths:
  /api/config:
    get:
      summary: Get global configuration
      tags: [Config]
      responses:
        '200': { description: OK }
    put:
      summary: Update global configuration
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Config'
      responses:
        '200': { description: Updated }

  /api/servers:
    get:
      summary: List remote servers
      tags: [Servers]
      responses:
        '200': { description: OK }
    post:
      summary: Add remote server
      tags: [Servers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerInput'
      responses:
        '200': { description: Created }

  /api/containers:
    get:
      summary: List agents
      description: Returns paginated list of agents. Use ?all=true for full list.
      tags: [Agents]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: show_deleted
          in: query
          schema: { type: boolean }
        - name: all
          in: query
          description: Return all items without pagination
          schema: { type: boolean }
      responses:
        '200': { description: OK }
    post:
      summary: Provision a new agent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                server_id: { type: string }
                template_id: { type: string }
      responses:
        '200': { description: Provisioning started }

  /api/containers/{id}:
    delete:
      summary: Delete an agent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  /api/containers/{id}/start:
    post:
      summary: Start an agent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  /api/containers/{id}/stop:
    post:
      summary: Stop an agent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  /api/containers/{id}/refresh:
    post:
      summary: Refresh agent status from remote
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  /api/caddy/download:
    get:
      summary: Download Caddyfile
      tags: [Config]
      responses:
        '200':
          description: Caddyfile download
          content:
            text/plain:
              schema:
                type: string
                format: binary

  /api/templates:
    get:
      summary: List all templates
      tags: [Templates]
      responses:
        '200': { description: OK }
    post:
      summary: Create or update a template
      tags: [Templates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
                display_name: { type: string }
                description: { type: string }
                enabled: { type: boolean }
                metadata: { type: string }
      responses:
        '200': { description: OK }

  /api/bastion:
    get:
      summary: List all Bastion IDs
      description: Returns all Bastion IDs with their assigned agents (No pagination).
      tags: [Bastion]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    createdAt: { type: string, format: date-time }
                    updatedAt: { type: string, format: date-time }
                    containers:
                      type: array
                      items:
                        type: object
                        properties:
                          id: { type: string }
                          domain_name: { type: string }
                          container_name: { type: string }
                          api_token: { type: string }
                          slug: { type: string }
                          status: { type: string }
    post:
      summary: Create a new Bastion ID
      tags: [Bastion]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
      responses:
        '200': { description: Created }

  /api/bastion/{id}:
    delete:
      summary: Delete a Bastion ID
      tags: [Bastion]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  /api/bastion/{id}/assign:
    post:
      summary: Assign or unassign an agent to a Bastion ID
      tags: [Bastion]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                container_id: { type: string }
                unassign: { type: boolean }
      responses:
        '200': { description: OK }

  /api/bastion/connections:
    get:
      summary: Get all active Bastion ID to Agent mappings
      description: Used by sync scripts on the bastion server. Returns multiple agents per ID.
      tags: [Bastion]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: object
                    properties:
                      container_id: { type: string }
                      slug: { type: string }
                      domain: { type: string }
                      api_token: { type: string }
                      status: { type: string }

components:
  schemas:
    Config:
      type: object
      properties:
        base_domain: { type: string }
        cloudflare_api_key: { type: string }
        cloudflare_zone_id: { type: string }
        target_cname: { type: string }
        caddyfile_path: { type: string }
    ServerInput:
      type: object
      required: [remote_ip, port, api_token]
      properties:
        remote_ip: { type: string }
        port: { type: number }
        api_token: { type: string }
        max_agents: { type: number }